gha for gh-pages and hugo (#3)
This should create a new GitHub action for `push` triggers.  The target branch is `main`.  Don't think this merge will trigger the GHA since at present, there is no GHA for `push` (or for any trigger for that matter).

But later, if I push to `main` (via merge of a PR), the GHA (github pages) should run.